requirements/teste.txt
-r base.txt
asynctest
pytest-cov
pytest-deadfixtures
pytest-django
pytest-env
pytest-pythonpath
pytest
